+++ ptai-cli-plugin/src/test/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/BaseCliIT.java
@@ -12,11 +12,14 @@ import java.nio.file.Path;
 
 public abstract class BaseCliIT extends BaseClientIT {
     protected Path PEM;
+    protected Path DUMMY;
 
     @SneakyThrows
     @BeforeEach
     public void pre() {
         PEM = Files.createTempFile(TEMP_FOLDER, "ptai-", "-ca");
         FileUtils.write(PEM.toFile(), BaseIT.CA, StandardCharsets.UTF_8);
+        DUMMY = Files.createTempFile(TEMP_FOLDER, "ptai-", "-ca");
+        FileUtils.write(DUMMY.toFile(), BaseIT.DUMMY, StandardCharsets.UTF_8);
     }
 }